+++ isis/src/mex/apps/hrsc2isis/hrsc2isis.h
@@ -1,0 +1,10 @@
+#ifndef hrsc2isis_h
+#define hrsc2isis_h
+
+#include "UserInterface.h"
+
+namespace Isis {
+  extern void hrsc2isis(UserInterface &ui);
+}
+
+#endif
