--- isis/src/mex/apps/hrsc2isis/tsts/phobos/Makefile
@@ -1,6 +1,0 @@
-APPNAME = hrsc2isis
-
-include $(ISISROOT)/make/isismake.tsts
-
-commands:
-	$(APPNAME) FROM=$(INPUT)/h7926_0009_s22.img TO=$(OUTPUT)/truth.cub > /dev/null;
