--- isis/src/mex/apps/hrsc2isis/tsts/srcImage/Makefile
@@ -1,9 +1,0 @@
-APPNAME =hrsc2isis
-
-include $(ISISROOT)/make/isismake.tsts
-
-commands:
-	$(APPNAME) FROM=$(INPUT)/h2862_0006_sr2.img \
-	           TO=$(OUTPUT)/h2862_0006_sr2.cub \
-	           > /dev/null
-
